lib/command: rework dispatching
Rework command dispatching to no longer create intermediary instances of
commands and then dispatch to them. This means that commands are now
unit structs, and the dispatch method is an associated method now.
Additionally, simplify the creation of requests into only a single
`Request` type.

The `CommandInfo` type that the command protocol returned is gone, and
now it just returns a `Request`. This same request can be used to
dispatch to commands.

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>