engine/command: use more efficient responses
When dispatching a command, instead of having the return type include a
new Vec containing the response, have commands write to a provided
mutably borrowed Vec. This allows consumers of the engine to re-use a
Vec, avoiding the allocation of a large number of needless Vecs.

This is the benchmark to perform an increment command before this
change:

test bench_increment ... bench:          85 ns/iter (+/- 7)

And after:

test bench_increment ... bench:          76 ns/iter (+/- 4)

Closes #2.

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>