client/backend/server: update response parsing
Update the response parsing to be up-to-date with the current response
scheme. This new method no longer uses archaic response formats based on
command and messages are no longer delimited by newlines.

Responses are now structured, and messages denote their length
dynamically based on the response type. For example, the first byte of a
response is the response type. If it's an integer, then following are 8
additional bytes in the message. This is the delimiter. The TCP reading
code now dynamically reads bytes of the stream based on what a context
parser provided by the `engine` instructs.

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>