client/cli: add 'stats' command
Add support for the Stats command. In the client, this is exposed as a
simple request struct that resolves to a 'StatsData' model struct, which
has methods for retrieving values of all of the metrics.

In the CLI, the stats command outputs all of the statistics as a list
delimited by newlines, like:

```
> stats
Commands successful: 15
Commands errored: 3
Sessions started: 1
Sessions ended: 0
```

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>