engine: add 'rename' command
Add a 'rename' command to the engine, supporting it in the client and
CLI.

The rename command renames a key from A to B, but only if B isn't
already taken at time of validation.

If the 'from' key isn't provided, then a `DispatchError::KeyRetrieval`
is returned.

If the `to` key isn't provided, then a
`DispatchError::ArgumentRetrieval` is returned.

If a key doesn't exist with the 'from' name, then a
`DispatchError::KeyRetrieval` is returned.

If a key doesn't exist with the 'to' name, then a
`DispatchError::PreconditionFailed` error is returned.

Using the CLI looks like this:

```
> rename
The key to rename is required.
> rename foo
The destination key name is required.
> rename foo bar
Dispatch error: key "foo" doesn't exist.
> set bar:int 1234
1
> rename foo bar
Dispatch error: the destination key, "bar", already exists.
> rename foo baz
baz
```

In the last command, the key is successfully renamed, and the new key
name is returned to confirm this.

Closes issue #9.

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>