engine: add 'delete' command
Add a 'delete' command to the engine, supporting it in the client and
CLI.

The delete command deletes a single key based on its name, but only if
the key exists.

If a key type is provided, then a `DispatchError::UnexpectedKeyType` is
returned.

If the key isn't provided, then a `DispatchError::KeyRetrieval` is
returned.

If the key is provided but it doesn't exist in Hop, then
`DispatchError::PreconditionFailed` is returned.

Using the CLI looks like this:

```
> delete
The key to delete is required.
> delete foo
Dispatch error: the key "foo" doesn't exist.
> set foo:bool true
true
> delete foo
foo
> delete foo
Dispatch error: the key "foo" doesn't exist.
```

When a key is successfully deleted, the key is returned as confirmation.

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>