engine: add 'exists' command (#20)
Add an 'exists' command to the engine, supporting it in the client and
CLI.

The exists command checks if one or more keys exist. If all exist, then
`true` is returned. If at least one does not exist, then `false` is
returned.

If a key type is provided, then a `DispatchError::UnexpectedKeyType` is
returned.

If no arguments are provided, then a `DispatchError::ArgumentRetrieval`
is returned.

Using the CLI looks like this:

```
> exists
At least one key is required.
> exists foo
false
> set foo 123
123
> exists foo
true
```

Closes issue #19.

Merged-by: Vivian Hellyer <vivian@hellyer.dev>
Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>