engine/response/context: don't reset on every iter
The response context was indvertently resetting the state on every
iteration, which was essentially throwing away a lot of a message's
processing on every iteration. Instead, only reset the state when a
message has completed parsing.

Take a scenario where a 15-element list of 100 byte elements is
transmitted over network. To parse this, a consumer will read the first
byte, which will be feed to the parser. The parser says it's a list and
to get 2 more bytes, so the consumer reads 2 more bytes, passes it to
the parser, and so on. For this list this will result in 62 reads to
completion. By not resetting the state, the time to parse this list
(minus networking) is reduced from:

```
test bench_response_list    ... bench:      13,033 ns/iter (+/- 1,666)
```

to:

```
test bench_response_list    ... bench:       3,681 ns/iter (+/- 184)
```

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>