cli: add wasm32-unknown-unknown support
Add support for the CLI by splitting out the processing work from the
I/O work. This now provides a `hop-cli` library that takes a string
input, parses it, runs the command with the provided client, and gives
back a clean human-readable string of the response or the error.

The binary CLI is now a light wrapper over the library that just loops
over input.

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>