engine: add 'is' command
Add an 'is' command, which can be used to check that one or more keys is
of a certain specified key type. The command returns a boolean. If
*at least* one of the keys does not exist or is not of the specified
type, then `false` is returned. This is short-circuiting.

A potential case is where you have a key "foo" that you *think* is a
string, but you want to verify this, then you'd use the `is` command
with a key type of `String` and an argument of "foo".

If a key type isn't specified, then a `DispatchError::KeyTypeRequired`
error is returned.

If no arguments are provided, then a `DispatchError::ArgumentRetrieval`
error is returned.

For example, the command can be visualised by the CLI:

```
> set:str foo the value
the value
> is:str foo
true
> is:str foo bar
false

```

And also by the client:

```rust
use hop::{Client, KeyType};

let client = Client::memory();

client.set("foo").int(123).await?;

assert!(client.is(KeyType::Integer).key("foo").await?);
```

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>