cli: add key type parsing (#35)
Add key type parsing to the CLI, allowing for specifying and using
commands that require or allow one. This also indirectly fixes argument
and key parsing. All arguments would simply be the raw string bytes in
arguments. Now, if a command accepts one key and multiple arguments, the
key will be sent as raw bytes in the arguments and the multiple
arguments would be correctly parsed to the key type (such as an
integer).

This also adds `is` support to the CLI, which was used for testing. This
now allows commands like this:

```
> increment:int foo
1
> is:int foo
true
> is:int foo bar
false
```

Merged-by: Vivian Hellyer <vivian@hellyer.dev>
Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>