engine/command: add key type invalid dispatch error
The `DispatchError::WrongType` error was being used for both when the
key type was not supported by a command and when a key's type was
different than the requested key type operation.

To fix this, split the variant into `KeyTypeInvalid` for the former case
and `KeyTypeDifferent` for the latter case. The new variant has error
number 7.

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>