engine: add 'type' command
Add the `type` command to the engine, which retrieves the type of a key.
The type is in the form of an integer reply, which can be converted to a
key type as key types have stable integer representations.

If a key is not specified, than a `DispatchError::KeyUnspecified` error
is returned.

If a key type is specified, then a `DispatchError::KeyTypeUnexpected`
error is returned.

If a key does not exist, then a `DispatchError::KeyNonexistent` error is
returned.

CLI example:

```
> set foo:int 123
123
> type foo
int
> set foo:str this is a utf-8 valid string
this is a utf-8 valid string
> type foo
str
```

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>