engine: add 'get' command
Add the `get` command, which retrieves the value of a key. A key type
can be specified, which will only get the key if it is of the specified
key type.

Returns a `DispatchError::KeyUnspecified` if the key is not specified.
Returns a `DispatchError::KeyNonexistent` if the key does not exist.
Returns a `DispatchError::KeyTypeDifferent` if the specified key type is
different than the key's type.

CLI example:

```
> set:int foo 123
123
> get foo
123
> get:int foo
123
> set:bool bar true
true
> get:bool bar
true
> get bar
true
```

Closes #7.

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>