engine/command/request: add a builder
Add a RequestBuilder struct for constructing requests. This is useful
because it simplifies code that constructs requests - specifically the
arguments of requests - and ensures that not too many arguments or any
empty arguments are in a request.

The `Request::new` and `Request::new_with_type` methods have been
removed in favour of the builder.

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>