client: add no_std support (#53)
Add `no_std` support in the client. There's now one feature, `std`,
which is always available. If this is enabled, then it will add
`std::error::Error` support for errors. If it is disabled, then `no_std`
support will be enabled.

Additionally, there is now a `tokio` feature which enables server
backend support. `tokio` requires the `std` feature to be enabled.

Merged-by: Vivian Hellyer <vivian@hellyer.dev>
Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>