engine: add state::key_type method
Add a method to the State, `key_type`, which retrieves the key type of a
key if it exists. This simplifies working around a key when you only
need the type of it, and reduces the time that a reference to the pair
in the state is held.

This also simplifies a lot of code, and fixes some code that was
needlessly creating keys that weren't being assigned a (non-empty)
value.

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>